+++ lib/operation_contract_execute.go
@@ -6,7 +6,9 @@ import (
 
 	"github.com/stellar/go/keypair"
 
+	"boscoin.io/sebak/lib/block"
 	sebakcommon "boscoin.io/sebak/lib/common"
+	"boscoin.io/sebak/lib/contract"
 	"boscoin.io/sebak/lib/contract/payload"
 	"boscoin.io/sebak/lib/error"
 	"boscoin.io/sebak/lib/storage"
@@ -63,17 +65,17 @@ func (o OperationBodyContractExecute) GetAmount() sebakcommon.Amount {
 }
 
 func FinishOperationBodyContractExecute(st *sebakstorage.LevelDBBackend, tx Transaction, op Operation) (err error) {
-	var baSource, baTarget *BlockAccount
-	if baSource, err = GetBlockAccount(st, tx.B.Source); err != nil {
+	var baSource, baTarget *block.BlockAccount
+	if baSource, err = block.GetBlockAccount(st, tx.B.Source); err != nil {
 		err = sebakerror.ErrorBlockAccountDoesNotExists
 		return
 	}
-	if baTarget, err = GetBlockAccount(st, op.B.TargetAddress()); err != nil {
+	if baTarget, err = block.GetBlockAccount(st, op.B.TargetAddress()); err != nil {
 		err = sebakerror.ErrorBlockAccountDoesNotExists
 		return
 	}
 
-	ctx := NewContractContext(baSource, st) // st as statedb
+	ctx := contract.NewContractContext(baSource, st) // st as statedb
 
 	exCode := &payload.ExecCode{
 		ContractAddress: baTarget.Address,
@@ -81,6 +83,6 @@ func FinishOperationBodyContractExecute(st *sebakstorage.LevelDBBackend, tx Tran
 		Args:            op.B.(OperationBodyContractExecute).Args,
 	}
 
-	_, err = ExecuteContract(ctx, exCode) //TODO: Where to pass the return value?
+	_, err = contract.ExecuteContract(ctx, exCode) //TODO: Where to pass the return value?
 	return
 }
