--- lib/state_object.go
@@ -1,13 +1,0 @@
-package sebak
-
-type DirtyStatus byte
-
-const (
-	StateObjectChanged DirtyStatus = iota
-	StateObjectDeleted
-)
-
-type StateObject struct {
-	Value interface{}
-	State DirtyStatus
-}
